chore(devcontainer): improve Makefile devcontainer support

+++ Makefile
@@ -6,37 +6,54 @@ SHELL := /bin/bash
 
 PROJECT_NAME := $(shell basename "$$(pwd)")
 DOTNET_IMAGE := mcr.microsoft.com/dotnet/sdk:8.0-jammy
-GOLANG_IMAGE := golang:1.24-alpine3.22
-PROTOC_IMAGE := rvolosatovs/protoc:4.1.0
+GOLANG_IMAGE := golang:1.24-alpine3.21
+PROTOC_IMAGE := proto-builder
 
+IS_INSIDE_DEVCONTAINER := $(REMOTE_CONTAINERS)
 BUILD_CACHE_VOLUME := $(shell echo '$(PROJECT_NAME)' | sed 's/[^a-zA-Z0-9_-]//g')-build-cache
 
 .PHONY: build
 
 build: build_server build_gateway
 
-proto:
-	docker run -t --rm \
-		-u $$(id -u):$$(id -g) \
-		-v $$(pwd):/data \
-		-w /data \
-		--entrypoint /bin/bash \
-		${PROTOC_IMAGE} \
-		proto.sh
+proto_image:
+ifneq ($(IS_INSIDE_DEVCONTAINER),true)
+	docker build --target proto-builder -t $(PROTOC_IMAGE) .
+endif
 
-build_server: prepare_build_cache
+proto: proto_image
+ifneq ($(IS_INSIDE_DEVCONTAINER),true)
+	docker run --tty --rm --user $$(id -u):$$(id -g) \
+		--volume $$(pwd):/build \
+		--workdir /build \
+		--entrypoint /bin/bash \
+		$(PROTOC_IMAGE) \
+		proto.sh
+else
+	./proto.sh
+endif
+
+build_server: proto prepare_build_cache
+ifneq ($(IS_INSIDE_DEVCONTAINER),true)
 	docker run -t --rm -u $$(id -u):$$(id -g) \
 		-e HOME="/tmp/build-cache/dotnet/cache" \
 		-e DOTNET_CLI_HOME="/tmp/build-cache/dotnet/cache" \
 		-e DOTNET_SKIP_WORKLOAD_INTEGRITY_CHECK=1 \
-		-v $(BUILD_CACHE_VOLUME):/tmp/build-cache \
 		-v $$(pwd):/data \
+		-v $(BUILD_CACHE_VOLUME):/tmp/build-cache \
 		-w /data/src \
 		${DOTNET_IMAGE} \
 		dotnet build
+else
+	cd src && dotnet build
+endif
 
-
-build_gateway: proto prepare_build_cache
+build_gateway: proto
+ifneq ($(IS_INSIDE_DEVCONTAINER),true)
+	docker run -t --rm \
+			-v $(BUILD_CACHE_VOLUME):/tmp/build-cache \
+			$(GOLANG_IMAGE) \
+			chown $$(id -u):$$(id -g) /tmp/build-cache		# Fix /tmp/build-cache folder owned by root
 	docker run -t --rm -u $$(id -u):$$(id -g) \
 			-e GOCACHE=/tmp/build-cache/go/cache \
 			-e GOMODCACHE=/tmp/build-cache/go/modcache \
@@ -45,8 +62,12 @@ build_gateway: proto prepare_build_cache
 			-w /data/gateway \
 			${GOLANG_IMAGE} \
 			go build -modcacherw -o grpc_gateway
+else
+	cd gateway && go build -modcacherw -o grpc_gateway
+endif
 
 run_server: prepare_build_cache
+ifneq ($(IS_INSIDE_DEVCONTAINER),true)
 	docker run --rm -it -u $$(id -u):$$(id -g) \
 		-e HOME="/tmp/build-cache/dotnet/cache" \
 		-e DOTNET_CLI_HOME="/tmp/build-cache/dotnet/cache" \
@@ -59,8 +80,16 @@ run_server: prepare_build_cache
 		-p 8080:8080 \
 		${DOTNET_IMAGE} \
 		dotnet run
+else
+	cd src/AccelByte.Extend.Vivox.Authentication.Server && dotnet run
+endif
 
-run_gateway: proto prepare_build_cache
+run_gateway: proto
+ifneq ($(IS_INSIDE_DEVCONTAINER),true)
+	docker run -t --rm \
+			-v $(BUILD_CACHE_VOLUME):/tmp/build-cache \
+			$(GOLANG_IMAGE) \
+			chown $$(id -u):$$(id -g) /tmp/build-cache		# Fix /tmp/build-cache folder owned by root
 	docker run -it --rm -u $$(id -u):$$(id -g) \
 			-e GOCACHE=/tmp/build-cache/go/cache \
 			-e GOMODCACHE=/tmp/build-cache/go/modcache \
@@ -72,9 +101,14 @@ run_gateway: proto prepare_build_cache
 			--add-host host.docker.internal:host-gateway \
 			${GOLANG_IMAGE} \
 			go run main.go --grpc-addr host.docker.internal:6565
+else
+	cd gateway && go run main.go --grpc-addr localhost:6565
+endif
 
 prepare_build_cache:
+ifneq ($(IS_INSIDE_DEVCONTAINER),true)
 	docker run -t --rm \
 			-v $(BUILD_CACHE_VOLUME):/tmp/build-cache \
 			busybox:1.37.0 \
 			chown $$(id -u):$$(id -g) /tmp/build-cache		# Fix /tmp/build-cache folder owned by root
+endif
